+++ src/everest/everest_storage.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import logging
-import os
 import traceback
 from pathlib import Path
 from typing import Any, cast
@@ -541,208 +540,3 @@ class EverestStorage:
             }
 
         return batch_dataframes
-
-    def export_dataframes(
-        self,
-    ) -> tuple[pl.DataFrame, pl.DataFrame, pl.DataFrame]:
-        batch_dfs_to_join = {}  # type: ignore
-        realization_dfs_to_join = {}  # type: ignore
-        perturbation_dfs_to_join = {}  # type: ignore
-
-        batch_ids = [b.batch_id for b in self.experiment.batches]
-        all_controls = self.control_names
-
-        def _try_append_df(
-            batch_id: int,
-            df: pl.DataFrame | None,
-            target: dict[str, list[pl.DataFrame]],
-        ) -> None:
-            if df is not None:
-                if batch_id not in target:  # type: ignore
-                    target[batch.batch_id] = []  # type: ignore
-                target[batch_id].append(df)  # type: ignore
-
-        def try_append_batch_dfs(batch_id: int, *dfs: pl.DataFrame | None) -> None:
-            for df_ in dfs:
-                _try_append_df(batch_id, df_, batch_dfs_to_join)
-
-        def try_append_realization_dfs(
-            batch_id: int, *dfs: pl.DataFrame | None
-        ) -> None:
-            for df_ in dfs:
-                _try_append_df(batch_id, df_, realization_dfs_to_join)
-
-        def try_append_perturbation_dfs(
-            batch_id: int, *dfs: pl.DataFrame | None
-        ) -> None:
-            for df_ in dfs:
-                _try_append_df(batch_id, df_, perturbation_dfs_to_join)
-
-        def pivot_gradient(df: pl.DataFrame) -> pl.DataFrame:
-            pivoted_ = df.pivot(on="control_name", index="batch_id", separator=" wrt ")
-            return pivoted_.rename(
-                {
-                    col: f"grad({col})"
-                    for col in pivoted_.columns
-                    if col != "batch_id" and col not in all_controls
-                }
-            )
-
-        for batch in self.experiment.batches:
-            if not batch.has_data:
-                continue
-
-            try_append_perturbation_dfs(
-                batch.batch_id,
-                batch.perturbation_objectives,
-                batch.perturbation_constraints,
-            )
-
-            try_append_realization_dfs(
-                batch.batch_id,
-                batch.realization_objectives,
-                batch.realization_controls,
-                batch.realization_constraints,
-            )
-
-            if batch.batch_objective_gradient is not None:
-                try_append_batch_dfs(
-                    batch.batch_id, pivot_gradient(batch.batch_objective_gradient)
-                )
-
-            if batch.batch_constraint_gradient is not None:
-                try_append_batch_dfs(
-                    batch.batch_id,
-                    pivot_gradient(batch.batch_constraint_gradient),
-                )
-
-            try_append_batch_dfs(
-                batch.batch_id,
-                batch.batch_objectives,
-                batch.batch_constraints,
-            )
-
-        def _join_by_batch(
-            dfs: dict[int, list[pl.DataFrame]], on: list[str]
-        ) -> list[pl.DataFrame]:
-            """
-            Creates one dataframe per batch, with one column per input/output,
-            including control, objective, constraint, gradient value.
-            """
-            dfs_to_concat_ = []
-            for batch_id in batch_ids:
-                if batch_id not in dfs:
-                    continue
-
-                batch_df_ = dfs[batch_id][0]
-                for bdf_ in dfs[batch_id][1:]:
-                    if set(all_controls).issubset(set(bdf_.columns)) and set(
-                        all_controls
-                    ).issubset(set(batch_df_.columns)):
-                        bdf_ = bdf_.drop(all_controls)
-
-                    batch_df_ = batch_df_.join(
-                        bdf_,
-                        on=on,
-                    )
-
-                dfs_to_concat_.append(batch_df_)
-
-            return dfs_to_concat_
-
-        batch_dfs_to_concat = _join_by_batch(batch_dfs_to_join, on=["batch_id"])
-        batch_df = pl.concat(batch_dfs_to_concat, how="diagonal")
-
-        realization_dfs_to_concat = _join_by_batch(
-            realization_dfs_to_join, on=["batch_id", "realization", "simulation_id"]
-        )
-        realization_df = pl.concat(realization_dfs_to_concat, how="diagonal")
-
-        perturbation_dfs_to_concat = _join_by_batch(
-            perturbation_dfs_to_join, on=["batch_id", "realization", "perturbation"]
-        )
-        if perturbation_dfs_to_concat:
-            # Perturbations exists, proceed as normal
-            perturbation_df = pl.concat(perturbation_dfs_to_concat, how="diagonal")
-            pert_real_df = pl.concat([realization_df, perturbation_df], how="diagonal")
-        else:
-            # Discrete methods never have perturbations,
-            # append an empty (i.e., null) column
-            pert_real_df = realization_df.with_columns(
-                pl.lit(None).alias("perturbation")
-            )
-
-        pert_real_df = pert_real_df.select(
-            "batch_id",
-            "realization",
-            "perturbation",
-            *list(
-                set(pert_real_df.columns) - {"batch_id", "realization", "perturbation"}
-            ),
-        )
-
-        # Avoid name collisions when joining with simulations
-        batch_df_renamed = batch_df.rename(
-            {
-                col: f"batch_{col}"
-                for col in batch_df.columns
-                if col != "batch_id" and not col.startswith("grad")
-            }
-        )
-        combined_df = pert_real_df.join(
-            batch_df_renamed, on="batch_id", how="full", coalesce=True
-        )
-
-        def _sort_df(df: pl.DataFrame, index: list[str]) -> pl.DataFrame:
-            sorted_cols = index + sorted(set(df.columns) - set(index))
-            df_ = df.select(sorted_cols).sort(by=index)
-            return df_
-
-        return (
-            _sort_df(
-                combined_df,
-                ["batch_id", "realization", "simulation_id", "perturbation"],
-            ),
-            _sort_df(
-                pert_real_df,
-                [
-                    "batch_id",
-                    "realization",
-                    "perturbation",
-                    "simulation_id",
-                ],
-            ),
-            _sort_df(batch_df, ["batch_id", "total_objective_value"]),
-        )
-
-    def export_everest_opt_results_to_csv(self) -> Path:
-        batches_with_data = ",".join(
-            {str(b.batch_id) for b in self.experiment.batches if b.has_data}
-        )
-        full_path = (
-            self.experiment._path
-            / f"experiment_results_batches::{','.join(batches_with_data)}.csv"
-        )
-
-        # Find old csv to delete
-        existing_csv = next(
-            (
-                Path(f)
-                for f in os.listdir(self.experiment._path)
-                if f.startswith("experiment_results_batches::")
-            ),
-            None,
-        )
-
-        if (
-            existing_csv is not None
-            and existing_csv.exists()
-            and (self.experiment._path / existing_csv) != full_path
-        ):
-            # New batches are added -> overwrite existing csv
-            os.remove(existing_csv)
-
-        if not os.path.exists(full_path):
-            combined_df, _, _ = self.export_dataframes()
-            combined_df.write_csv(full_path)
-        return full_path
